Refactored ACL type to use "action" parameter instead of nonsensical "ensure" property
The type as written abused the syntax of the "ensure" property. Since nothing is being created or deleted, and since the "ensure" convention doesn't match the actions available for ACLs very well, I replaced it with the "action" parameter. In addition, I refactored the getter and setter for the "permission" property, cleaned up the logic surrounding set/unset/purge of ACLs, and added actions to set ACLs destructively or non-destructively (action => set or action => exact, respectively.)